requirements.txt
yfinance>=0.2.28
pandas>=2.0
numpy>=1.24
backtesting>=0.3.3
riskfolio-lib>=4.0
matplotlib>=3.7
alpaca-trade-api>=3.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
pytz>=2023.3
